+++ cmd/harp/internal/cmd/bundle_patch.go
@@ -21,6 +21,7 @@ import (
 	"github.com/spf13/cobra"
 	"go.uber.org/zap"
 
+	"github.com/elastic/harp/pkg/bundle/patch"
 	"github.com/elastic/harp/pkg/sdk/cmdutil"
 	"github.com/elastic/harp/pkg/sdk/log"
 	"github.com/elastic/harp/pkg/tasks/bundle"
@@ -28,17 +29,22 @@ import (
 )
 
 // -----------------------------------------------------------------------------
+type bundlePatchParams struct {
+	inputPath         string
+	outputPath        string
+	patchPath         string
+	valueFiles        []string
+	values            []string
+	stringValues      []string
+	fileValues        []string
+	stopAtRuleIndex   int
+	stopAtRuleID      string
+	ignoreRuleIDs     []string
+	ignoreRuleIndexes []int
+}
 
 var bundlePatchCmd = func() *cobra.Command {
-	var (
-		inputPath    string
-		outputPath   string
-		patchPath    string
-		valueFiles   []string
-		values       []string
-		stringValues []string
-		fileValues   []string
-	)
+	params := &bundlePatchParams{}
 
 	cmd := &cobra.Command{
 		Use:   "patch",
@@ -50,22 +56,31 @@ var bundlePatchCmd = func() *cobra.Command {
 
 			// Load values
 			valueOpts := tplcmdutil.ValueOptions{
-				ValueFiles:   valueFiles,
-				Values:       values,
-				StringValues: stringValues,
-				FileValues:   fileValues,
+				ValueFiles:   params.valueFiles,
+				Values:       params.values,
+				StringValues: params.stringValues,
+				FileValues:   params.fileValues,
 			}
 			values, err := valueOpts.MergeValues()
 			if err != nil {
 				log.For(ctx).Fatal("unable to process values", zap.Error(err))
 			}
 
+			// Prepare patch options.
+			opts := []patch.OptionFunc{
+				patch.WithStopAtRuleID(params.stopAtRuleID),
+				patch.WithStopAtRuleIndex(params.stopAtRuleIndex),
+				patch.WithIgnoreRuleIDs(params.ignoreRuleIDs...),
+				patch.WithIgnoreRuleIndexes(params.ignoreRuleIndexes...),
+			}
+
 			// Prepare task
 			t := &bundle.PatchTask{
-				ContainerReader: cmdutil.FileReader(inputPath),
-				PatchReader:     cmdutil.FileReader(patchPath),
-				OutputWriter:    cmdutil.FileWriter(outputPath),
+				ContainerReader: cmdutil.FileReader(params.inputPath),
+				PatchReader:     cmdutil.FileReader(params.patchPath),
+				OutputWriter:    cmdutil.FileWriter(params.outputPath),
 				Values:          values,
+				Options:         opts,
 			}
 
 			// Run the task
@@ -76,14 +91,18 @@ var bundlePatchCmd = func() *cobra.Command {
 	}
 
 	// Parameters
-	cmd.Flags().StringVar(&inputPath, "in", "-", "Container input ('-' for stdin or filename)")
-	cmd.Flags().StringVar(&outputPath, "out", "", "Container output ('-' for stdout or a filename)")
-	cmd.Flags().StringVar(&patchPath, "spec", "", "Patch specification path ('-' for stdin or filename)")
+	cmd.Flags().StringVar(&params.inputPath, "in", "-", "Container input ('-' for stdin or filename)")
+	cmd.Flags().StringVar(&params.outputPath, "out", "", "Container output ('-' for stdout or a filename)")
+	cmd.Flags().StringVar(&params.patchPath, "spec", "", "Patch specification path ('-' for stdin or filename)")
 	log.CheckErr("unable to mark 'spec' flag as required.", cmd.MarkFlagRequired("spec"))
-	cmd.Flags().StringArrayVar(&valueFiles, "values", []string{}, "Specifies value files to load")
-	cmd.Flags().StringArrayVar(&values, "set", []string{}, "Specifies value (k=v)")
-	cmd.Flags().StringArrayVar(&stringValues, "set-string", []string{}, "Specifies value (k=string)")
-	cmd.Flags().StringArrayVar(&fileValues, "set-file", []string{}, "Specifies value (k=filepath)")
+	cmd.Flags().StringArrayVar(&params.valueFiles, "values", []string{}, "Specifies value files to load")
+	cmd.Flags().StringArrayVar(&params.values, "set", []string{}, "Specifies value (k=v)")
+	cmd.Flags().StringArrayVar(&params.stringValues, "set-string", []string{}, "Specifies value (k=string)")
+	cmd.Flags().StringArrayVar(&params.fileValues, "set-file", []string{}, "Specifies value (k=filepath)")
+	cmd.Flags().StringVar(&params.stopAtRuleID, "stop-at-rule-id", "", "Stop patch evaluation before the given rule ID")
+	cmd.Flags().IntVar(&params.stopAtRuleIndex, "stop-at-rule-index", -1, "Stop patch evaluation before the given rule index (0 for first rule)")
+	cmd.Flags().StringArrayVar(&params.ignoreRuleIDs, "ignore-rule-id", []string{}, "List of Rule identifier to ignore during evaluation")
+	cmd.Flags().IntSliceVar(&params.ignoreRuleIndexes, "ignore-rule-index", []int{}, "List of Rule index to ignore during evaluation")
 
 	return cmd
 }