+++ pkg/bundle/patch/package_test.go
@@ -242,9 +242,10 @@ func mustLoadPatch(filePath string) *bundlev1.Patch {
 
 func TestApply(t *testing.T) {
 	type args struct {
-		spec   *bundlev1.Patch
-		b      *bundlev1.Bundle
-		values map[string]interface{}
+		spec    *bundlev1.Patch
+		b       *bundlev1.Bundle
+		values  map[string]interface{}
+		options []OptionFunc
 	}
 	tests := []struct {
 		name    string
@@ -452,6 +453,206 @@ func TestApply(t *testing.T) {
 			want: &bundlev1.Bundle{
 				Packages: []*bundlev1.Package{
 					{
+						Name: "application/another-created-package",
+						Annotations: map[string]string{
+							"package-creator":                       "true",
+							"patched":                               "true",
+							"secret-service.elstc.co/encryptionKey": "UcbPlrEJ9jZEQX06n8oMln_mCl3EU2zl2ZVc-obb7Dw=",
+						},
+						Secrets: &bundlev1.SecretChain{
+							Annotations: map[string]string{
+								"secret-service.elstc.co/encryptionKey": "DrZ-0yEA18iS7A4xaR_pd-relh9KMtTw2q11nBEJykg=",
+							},
+							Data: []*bundlev1.KV{
+								{
+									Key:   "key",
+									Type:  "string",
+									Value: []byte("0\n\x02\x01\x01\x13\x05value"),
+								},
+							},
+						},
+					},
+					{
+						Name: "application/created-package",
+						Annotations: map[string]string{
+							"package-creator":                       "true",
+							"patched":                               "true",
+							"secret-service.elstc.co/encryptionKey": "UcbPlrEJ9jZEQX06n8oMln_mCl3EU2zl2ZVc-obb7Dw=",
+						},
+						Secrets: &bundlev1.SecretChain{
+							Annotations: map[string]string{
+								"secret-service.elstc.co/encryptionKey": "DrZ-0yEA18iS7A4xaR_pd-relh9KMtTw2q11nBEJykg=",
+							},
+							Data: []*bundlev1.KV{
+								{
+									Key:   "key",
+									Type:  "string",
+									Value: []byte("0\n\x02\x01\x01\x13\x05value"),
+								},
+							},
+						},
+					},
+					{
+						Name: "secrets/application/component-2.yaml",
+					},
+				},
+			},
+		},
+		{
+			name: "add package - stop at 1",
+			args: args{
+				spec: mustLoadPatch("../../../test/fixtures/patch/valid/add-package.yaml"),
+				b: &bundlev1.Bundle{
+					Packages: []*bundlev1.Package{
+						{
+							Name: "secrets/application/component-2.yaml",
+						},
+					},
+				},
+				values: map[string]interface{}{},
+				options: []OptionFunc{
+					WithStopAtRuleIndex(1),
+				},
+			},
+			wantErr: false,
+			want: &bundlev1.Bundle{
+				Packages: []*bundlev1.Package{
+					{
+						Name: "application/created-package",
+						Annotations: map[string]string{
+							"package-creator":                       "true",
+							"patched":                               "true",
+							"secret-service.elstc.co/encryptionKey": "UcbPlrEJ9jZEQX06n8oMln_mCl3EU2zl2ZVc-obb7Dw=",
+						},
+						Secrets: &bundlev1.SecretChain{
+							Annotations: map[string]string{
+								"secret-service.elstc.co/encryptionKey": "DrZ-0yEA18iS7A4xaR_pd-relh9KMtTw2q11nBEJykg=",
+							},
+							Data: []*bundlev1.KV{
+								{
+									Key:   "key",
+									Type:  "string",
+									Value: []byte("0\n\x02\x01\x01\x13\x05value"),
+								},
+							},
+						},
+					},
+					{
+						Name: "secrets/application/component-2.yaml",
+					},
+				},
+			},
+		},
+		{
+			name: "add package - stop at 'another-package'",
+			args: args{
+				spec: mustLoadPatch("../../../test/fixtures/patch/valid/add-package.yaml"),
+				b: &bundlev1.Bundle{
+					Packages: []*bundlev1.Package{
+						{
+							Name: "secrets/application/component-2.yaml",
+						},
+					},
+				},
+				values: map[string]interface{}{},
+				options: []OptionFunc{
+					WithStopAtRuleID("another-package"),
+				},
+			},
+			wantErr: false,
+			want: &bundlev1.Bundle{
+				Packages: []*bundlev1.Package{
+					{
+						Name: "application/created-package",
+						Annotations: map[string]string{
+							"package-creator":                       "true",
+							"patched":                               "true",
+							"secret-service.elstc.co/encryptionKey": "UcbPlrEJ9jZEQX06n8oMln_mCl3EU2zl2ZVc-obb7Dw=",
+						},
+						Secrets: &bundlev1.SecretChain{
+							Annotations: map[string]string{
+								"secret-service.elstc.co/encryptionKey": "DrZ-0yEA18iS7A4xaR_pd-relh9KMtTw2q11nBEJykg=",
+							},
+							Data: []*bundlev1.KV{
+								{
+									Key:   "key",
+									Type:  "string",
+									Value: []byte("0\n\x02\x01\x01\x13\x05value"),
+								},
+							},
+						},
+					},
+					{
+						Name: "secrets/application/component-2.yaml",
+					},
+				},
+			},
+		},
+		{
+			name: "add package - ignore 0",
+			args: args{
+				spec: mustLoadPatch("../../../test/fixtures/patch/valid/add-package.yaml"),
+				b: &bundlev1.Bundle{
+					Packages: []*bundlev1.Package{
+						{
+							Name: "secrets/application/component-2.yaml",
+						},
+					},
+				},
+				values: map[string]interface{}{},
+				options: []OptionFunc{
+					WithIgnoreRuleIndexes(0),
+				},
+			},
+			wantErr: false,
+			want: &bundlev1.Bundle{
+				Packages: []*bundlev1.Package{
+					{
+						Name: "application/another-created-package",
+						Annotations: map[string]string{
+							"package-creator":                       "true",
+							"patched":                               "true",
+							"secret-service.elstc.co/encryptionKey": "UcbPlrEJ9jZEQX06n8oMln_mCl3EU2zl2ZVc-obb7Dw=",
+						},
+						Secrets: &bundlev1.SecretChain{
+							Annotations: map[string]string{
+								"secret-service.elstc.co/encryptionKey": "DrZ-0yEA18iS7A4xaR_pd-relh9KMtTw2q11nBEJykg=",
+							},
+							Data: []*bundlev1.KV{
+								{
+									Key:   "key",
+									Type:  "string",
+									Value: []byte("0\n\x02\x01\x01\x13\x05value"),
+								},
+							},
+						},
+					},
+					{
+						Name: "secrets/application/component-2.yaml",
+					},
+				},
+			},
+		},
+		{
+			name: "add package - ignore id",
+			args: args{
+				spec: mustLoadPatch("../../../test/fixtures/patch/valid/add-package.yaml"),
+				b: &bundlev1.Bundle{
+					Packages: []*bundlev1.Package{
+						{
+							Name: "secrets/application/component-2.yaml",
+						},
+					},
+				},
+				values: map[string]interface{}{},
+				options: []OptionFunc{
+					WithIgnoreRuleIDs("another-package"),
+				},
+			},
+			wantErr: false,
+			want: &bundlev1.Bundle{
+				Packages: []*bundlev1.Package{
+					{
 						Name: "application/created-package",
 						Annotations: map[string]string{
 							"package-creator":                       "true",
@@ -480,12 +681,12 @@ func TestApply(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			got, err := Apply(tt.args.spec, tt.args.b, tt.args.values)
+			got, err := Apply(tt.args.spec, tt.args.b, tt.args.values, tt.args.options...)
 			if (err != nil) != tt.wantErr {
 				t.Errorf("Apply() error = %v, wantErr %v", err, tt.wantErr)
 				return
 			}
-			if diff := cmp.Diff(got, tt.want, ignoreOpts...); diff != "" {
+			if diff := cmp.Diff(tt.want, got, ignoreOpts...); diff != "" {
 				t.Errorf("%q. Patch.Apply():\n-got/+want\ndiff %s", tt.name, diff)
 			}
 		})